requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
httpx
pydantic
requests
python-jose[cryptography]
bcrypt
passlib[bcrypt]